+++ app/src/main/java/com/mobile/agri10x/models/GetCreateBookingCartData.java
@@ -12,7 +12,7 @@ public class GetCreateBookingCartData {
     private List<GetCreateBookCartDataProduct> products = null;
     @SerializedName("subTotal")
     @Expose
-    private Integer subTotal;
+    private double subTotal;
 
     public List<GetCreateBookCartDataProduct> getProducts() {
         return products;
@@ -22,11 +22,11 @@ public class GetCreateBookingCartData {
         this.products = products;
     }
 
-    public Integer getSubTotal() {
+    public double getSubTotal() {
         return subTotal;
     }
 
-    public void setSubTotal(Integer subTotal) {
+    public void setSubTotal(double subTotal) {
         this.subTotal = subTotal;
     }
 }
