+++ app/src/main/java/com/mobile/agri10x/models/GetCreateCheckoutCartProductCheckout.java
@@ -18,7 +18,7 @@ public class GetCreateCheckoutCartProductCheckout {
     private String variety;
     @SerializedName("price")
     @Expose
-    private Integer price;
+    private double price;
     @SerializedName("weight")
     @Expose
     private Integer weight;
@@ -70,7 +70,7 @@ public class GetCreateCheckoutCartProductCheckout {
         this.variety = variety;
     }
 
-    public Integer getPrice() {
+    public double getPrice() {
         return price;
     }
